对gys表进行微调 完成进货表（Jhd）进货详情（Jhxq）库存（kc）的联合存储。 完成了新建进货单页面

+++ src/main/java/com/action/GysAction.java
@@ -30,7 +30,7 @@ public class GysAction {
 
     private Gys gys;
 
-    private Long gysbh;
+    private Integer gysbh;
 
 
     public void setExcel(File excel) {
@@ -53,7 +53,7 @@ public class GysAction {
         this.gys = gys;
     }
 
-    public void setGysbh(Long gysbh) {
+    public void setGysbh(Integer gysbh) {
         this.gysbh = gysbh;
     }
